+++ contracts/governance/Version.sol
@@ -17,7 +17,7 @@ contract Version is DBC, Owned {
         address owner;
         string name;
         string symbol;
-        uint8 decimals;
+        uint decimals;
         bool active;
         uint timestamp;
         bytes32 ipfsHash;
@@ -46,7 +46,7 @@ contract Version is DBC, Owned {
         address owner,
         string name,
         string symbol,
-        uint8 decimals,
+        uint decimals,
         bool active,
         uint id,
         address ofUniverse,
@@ -84,7 +84,7 @@ contract Version is DBC, Owned {
     function createVault(
         string withName,
         string withSymbol,
-        uint8 withDecimals,
+        uint withDecimals,
         address ofUniverse,
         address ofParticipation,
         address ofRiskMgmt,
